Changes from OpenXR 1.0.27

+++ src/conformance/framework/action_utils.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2019-2022, The Khronos Group Inc.
+// Copyright (c) 2019-2023, The Khronos Group Inc.
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -161,7 +161,7 @@ namespace Conformance
 
         *static_cast<XrCompositionLayerQuad*>(this) = {XR_TYPE_COMPOSITION_LAYER_QUAD};
         size.width = 1;
-        size.height = size.width * image->height / image->width;
+        size.height = size.width * static_cast<float>(image->height) / static_cast<float>(image->width);
         pose = XrPosef{{0, 0, 0, 1}, {0, 0, -1.5f}};
         subImage = m_compositionHelper.MakeDefaultSubImage(messageSwapchain);
         space = compositionSpace;
